+++ src/commands/clean.ts
@@ -1,7 +1,7 @@
 import { I_LIKE_JS } from "../constants.ts";
 import { CheckForPath } from "../functions/filesystem.ts";
 import { LogStuff } from "../functions/io.ts";
-import { GetAllProjects, GetProjectSettings, NameProject, UnderstandProjectSettings } from "../functions/projects.ts";
+import { GetAllProjects, GetProjectSettings, NameProject, SpotProject, UnderstandProjectSettings } from "../functions/projects.ts";
 import type { TheCleanerConstructedParams } from "./constructors/command.ts";
 import GenericErrorHandler from "../utils/error.ts";
 import { PerformCleaning, PerformHardCleanup, ResolveLockfiles, ShowReport, ValidateIntensity } from "./toolkit/cleaner.ts";
@@ -11,11 +11,14 @@ import { GetElapsedTime } from "../functions/date.ts";
 export type tRESULT = { path: string; status: string; elapsedTime: string };
 
 export default async function TheCleaner(params: TheCleanerConstructedParams) {
-    const { intensity, verbose, update, lint, prettify, destroy, commit } = params;
-    // start time
-    const now = new Date();
-
     try {
+        // params
+        const { verbose, update, lint, prettify, destroy, commit } = params.flags;
+        const { intensity, project } = params.parameters;
+
+        // start time
+        const startTime = new Date();
+
         // original path
         const originalLocation = Deno.cwd();
         const realIntensity: CleanerIntensity = await ValidateIntensity(intensity);
@@ -44,6 +47,35 @@ export default async function TheCleaner(params: TheCleanerConstructedParams) {
             verbose,
         );
 
+        if (project !== 0) {
+            const workingProject = await SpotProject(project);
+            if (!workingProject) {
+                throw new Error(`Project not found at ${project}... what's up?`);
+            }
+            Deno.chdir(workingProject);
+            const { doClean, doUpdate, doPrettify, doLint, doDestroy } = UnderstandProjectSettings.protection(
+                await GetProjectSettings(project),
+                {
+                    update,
+                    lint,
+                    destroy,
+                    prettify,
+                },
+            );
+            await PerformCleaning(
+                workingProject,
+                doUpdate,
+                doClean,
+                doLint,
+                doPrettify,
+                doDestroy,
+                commit,
+                realIntensity,
+            );
+            Deno.chdir(originalLocation);
+            return;
+        }
+
         const results: tRESULT[] = [];
 
         for (const project of projects) {
@@ -56,7 +88,7 @@ export default async function TheCleaner(params: TheCleanerConstructedParams) {
                 results.push({
                     path: project,
                     status: "Not found",
-                    elapsedTime: GetElapsedTime(now),
+                    elapsedTime: GetElapsedTime(startTime),
                 });
                 continue;
             }
@@ -102,7 +134,7 @@ export default async function TheCleaner(params: TheCleanerConstructedParams) {
                         results.push({
                             path: project,
                             status: "Too many lockfiles.",
-                            elapsedTime: GetElapsedTime(now),
+                            elapsedTime: GetElapsedTime(startTime),
                         });
                         continue;
                     }
@@ -115,7 +147,7 @@ export default async function TheCleaner(params: TheCleanerConstructedParams) {
                     results.push({
                         path: project,
                         status: "No lockfile.",
-                        elapsedTime: GetElapsedTime(now),
+                        elapsedTime: GetElapsedTime(startTime),
                     });
                     continue;
                 } else {
@@ -127,7 +159,7 @@ export default async function TheCleaner(params: TheCleanerConstructedParams) {
                     results.push({
                         path: project,
                         status: "No package.json.",
-                        elapsedTime: GetElapsedTime(now),
+                        elapsedTime: GetElapsedTime(startTime),
                     });
                     continue;
                 }
@@ -135,7 +167,7 @@ export default async function TheCleaner(params: TheCleanerConstructedParams) {
                 results.push({
                     path: project,
                     status: /* preliminaryStatus ? `Success # ${preliminaryStatus}` : */ "Success",
-                    elapsedTime: GetElapsedTime(now),
+                    elapsedTime: GetElapsedTime(startTime),
                 });
             } catch (e) {
                 await LogStuff(
@@ -146,7 +178,7 @@ export default async function TheCleaner(params: TheCleanerConstructedParams) {
                 results.push({
                     path: project,
                     status: "Failed",
-                    elapsedTime: GetElapsedTime(now),
+                    elapsedTime: GetElapsedTime(startTime),
                 });
                 continue;
             }